+++ config/kube.py
@@ -17,7 +17,7 @@ import inspect
 import shutil
 from urlparse import urlparse
 from swsssdk import ConfigDBConnector
-from config.main import run_command, AbbreviationGroup
+from common import run_command, AbbreviationGroup
 
 logger_identity = "kube_config"
 log_level = syslog.LOG_DEBUG
@@ -97,9 +97,9 @@ def _take_lock():
         lock_fd = open(LOCK_FILE, "w")
         fcntl.lockf(lock_fd, fcntl.LOCK_EX | fcntl.LOCK_NB)
         _log_info("Lock taken {}".format(LOCK_FILE))
-    except:
+    except IOError as e:
         lock_fd = None
-        _log_err("Lock {} failed: {}".format(LOCK_FILE, str(sys.exc_info())))
+        _log_err("Lock {} failed: {}".format(LOCK_FILE, str(e)))
     return lock_fd
 
 def _drop_cron():
@@ -235,9 +235,6 @@ def _do_join(server,insecure):
     except OSError as e:
         _do_exit("Download failed: {}".format(str(sys.exc_info()[0])))
 
-    except:
-        _do_exit("kube join failed: {}".format(str(sys.exc_info()[0])))
-
     _troubleshoot_tips()
 
 
@@ -332,7 +329,7 @@ kubernetes.add_command(server)
 @click.command()
 @click.argument('vip')
 def ip(vip):
-    """Specify a kubernetes ckuster VIP"""
+    """Specify a kubernetes cluster VIP"""
     if not netaddr.IPAddress(vip):
         click.echo('Invalid IP address %s' % vip)
         return
@@ -341,17 +338,17 @@ server.add_command(ip)
 
 #cmd kubernetes server insecure
 @click.command()
-@click.argument('option', type=click.Choice(["True", "False"]))
+@click.argument('option', type=click.Choice(["On", "Off"]))
 def insecure(option):
-    """Specify a kubernetes ckuster VIP access as insecure or not"""
+    """Specify a kubernetes cluster VIP access as insecure or not"""
     _update_kube_server('insecure', option)
 server.add_command(insecure)
 
 #cmd kubernetes server disable
 @click.command()
-@click.argument('option', type=click.Choice(["True", "False"]))
+@click.argument('option', type=click.Choice(["On", "Off"]))
 def disable(option):
-    """Specify a kubernetes ckuster VIP access is disabled or not"""
+    """Specify a kubernetes cluster VIP access is disabled or not"""
     _update_kube_server('disable', option)
 server.add_command(disable)
